+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog for WindowsUpdateSetting
 
+## v0.2.0
+
++ Updated manifest
++ Added functions to defer feature and quality updates
++ Added function set manage Windows Active hours
++ Modified `Suspend-WindowsUpdate` to specify a resume date.
++ Updated documentation
+
 ## v0.1.0
 
 + initial files